requirements.txt
opencv-python==4.12.0.88
numpy==2.2.6
pytesseract==0.3.13
python-docx==1.2.0
ultralytics==8.3.237
huggingface_hub
rembg==2.0.69
torch==2.9.1
torchvision==0.24.1
transformers==4.47.0
timm==0.9.2
# FastAPI & API dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
prometheus-client>=0.21.0
pdf2image>=1.17.0
# Production server
gunicorn>=23.0.0
